Bank test

+++ demo/src/main/kotlin/com/enihsyou/rpc/demo/Domians.kt
@@ -5,5 +5,5 @@ import java.math.BigDecimal
 data class User(
     val username: String,
     val password: String,
-    var money: BigDecimal
+    var money: Double
 )
